requirements.txt
# PyPDF2
# python-dotenv
# langchain
# groq
# pinecone
# langchain-text-splitters
# langchain-community
# sentence-transformers
# langchain_huggingface


ipykernel==7.1.0
ipython==9.6.0
ipywidgets==8.1.7
langchain==1.0.3
langchain-community==0.4.1
langchain-core==1.0.2
langchain-huggingface==1.0.0
langchain-text-splitters==1.0.0
pinecone==7.3.0
PyPDF2==3.0.1
python-dotenv==1.2.1
sentence-transformers==5.1.2
groq
langchain_huggingface
